refactor(x/http/client): Move the file directory
