+++ x/net/http/_demo/headers/headers.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/goplus/llgoexamples/x/http"
+	"github.com/goplus/llgoexamples/x/net/http"
 )
 
 func main() {
