+++ x/net/http/_demo/postform/postform.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"net/url"
 
-	"github.com/goplus/llgo/x/http"
+	"github.com/goplus/llgo/x/net/http"
 )
 
 func main() {
