+++ x/net/http/_demo/timeout/timeout.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"time"
 
-	"github.com/goplus/llgoexamples/x/http"
+	"github.com/goplus/llgoexamples/x/net/http"
 )
 
 func main() {
